Add code to compute XOR on files.
We add a bit of code to read a file and compute an XOR on its bytes.
This lets us verify that we know how to read a file in chunks and also
perform a computation on the data in it.

We also add a couple of files of test data which we can use when
computing checksums.