Change test data, and fix exposed bug.
Our test data was blocked at 64kB boundaries, which matched the read
block size. This meant that data which did not match this block size
would result in incorrect checksums being computed, as the entire blocks
were being fed into the digests whether or not there was valid data in
them. We fix this with slices, and replace our test data with data which
doesn't land on any block boundaries, even word boundaries.

Our tests are updated with the new digest values, of course.