Create alternate MD5 digest module.
Instead of computing the MD5 in the `digest` module, we move it into the
`md5` module. It is still computed in the background, but this leaves us
better able to play with how we do it. We expect to get clearer, tighter
code as a result.

The `main` code still uses the implementation in `digest`. We rely on
our tests in the `md5::tests` module to ensure that the implementation
in `md5` is working correctly.