Use a `Message` enum to commuicated with the background thread.
Instead of communicating with the background MD5 thread using only data
blocks, use a (private) `Message` enum. This will allow us to
communicate in a clearer fashion with the thread, exiting the loop with
a `Finish` message instead of knowing that we should exit when a zero
length data block is sent. This also prevents us from premature
termination should the user call `append` with a zero length data block.