Keep the MD5 background thread alive across multiple digests.
Instead of terminating the MD5 background thread after a single digest
is computed, we leave the thread running in an infinite loop. The digest
will be reset after each call to `result()`.

This should leave us with a slightly more efficient implementation (and
simpler code) when we tie it back into the `main` function.

We also add some test code to make sure the digests are correctly
computed when we have perform multiple digests.