Convert our `test_digests` constants to actual `Digest` values.
Now that we can perform equality tests on `Digest` values, it makes
sense to provide actual `MD5` and `SHA256` values instead of the digest
data in `test_digests`. This permits us to remove a large amount of test
code without reducing the coverage of the tests. In fact, they become
clearer with our ability to say things like:

    assert_eq!(digest, MD5_ZERO_EMPTY);