Implement a CRC32 `digest::Digest`.
Though we don't really expect to need it, we include a CRC32 for
completeness sake. The value is added to the `Digest` enum, and we
implement all the usual traits (`Eq`, `Debug`, `Display`) on it.

We also add a CRC32 `Generator`, and include a `crc32()` utility
function to create one.

`main()` gets rewritten to use the new CRC32 `Generator`.

Test CRC32 values are added to `digest::test_digests`.