Add `digest_file()` function to `main`.
The `digest_file()` function is used to compute a collection of
cryptographic digests on a file. The digests are computed in parallel
using the `Generator` structs passed back from `crc32()`, `md5()`, or
`sha256()` in `digest`. The exact digests will depend on the generators
passed into the function.

Tests are included which compute CRC32, MD5, and SHA256 values for an
empty file, and file filled with zeros, and a file filled with random
data. A negative test is also included for a file which doesn't exist.