Remove the `crc32()`, `md5()`, and `sha256()` functions from `main`.
We no longer use the `crc32()`, `md5()`, or `sha256()` functions in
`main`, as the `digest_file()` now does their work (and does it in
parallel), so we remove them. The corresponding tests are also removed.

In addition, we change the `use digest` statement to import the
`crc32()`, `md5()`, and `sha256()` functions from `digest` now that
there isn't a naming conflict.