Add cargo build script to locate OpenSSL library directory.
We add a `build.rs` script to locate the OpenSSL library on the system.
This will check in the usual library locations on a Unix system, plus
the Homebrew location on a Mac system.

If the OpenSSL library is found, then the library path is added to the
library search path for the linker. If it is not, then the build script
will exit with an error code and a failure message to that effect.