Rewrite the `Config` command line parser.
We completely rewrite the `Config` command line parser to take input
flags telling us which digests to compute on the files. We also add
support for `--help` and `-h` flags, detect the need to use `stdin` if
no files are specified, and include a default set of digests to use if
no digests are specified.

The functions in `main.rs` are updated to work with the revised `Config`
structure, but the logic there is left unchanged. Therefore, although
config gives us many new features we can't actually use them quite yet.