Add integration tests to check the operation of the binary.
We aren't able to check a number of functions in `main.rs` because we
can't capture `stdout` or `stderr` (not to mention being unable to write
to `stdin`). We cover them with integration tests in `checksum.rs` which
run the binary and check the `stdout` and `stderr` results against
expected values.

We remove the unimplemented tests from `main.rs` now that these are in
place.