Reduce the buffer size of our reads.
Rudimentary performance checking shows that we can reduce our read
buffer size to 256kB and only loose about 2% of our performance. We also
use about 30% of the memory we were using when we had a 1MB buffer, so I
feel this is a worthwhile tradeoff. Our memory usage on a 64-bit Mac is
sitting around 2.4MB.