Move `Digest` into `digest` module, and implement two test digests.
We re-export `digest::Digest` at the crate.

The `Count` digest algorithm just keeps a modulo 256 count of the number
of bytes passed to the digest. The `XOR` digest algorithm keeps a
running XOR of the bytes passed. Both are useful for testing purposes,
but are only used for internal module testing.

We also remove the `Null` digest from `background`, replacing it with
the test digests introduced in `digest`.